MWPW-154184 - moved pagination logic to PartnerCards
