+++ edsdme/blocks/announcements/announcements.js
@@ -32,11 +32,16 @@ export default async function init(el) {
     '{{filter-by}}': 'Filter by',
     '{{filters}}': 'Filters',
     '{{last-90-days}}': 'Last 90 days',
+    '{{next}}': 'Next',
+    '{{next-page}}': 'Next Page',
     '{{load-more}}': 'Load more',
     '{{no-results-description}}': 'Try checking your spelling or broadening your search.',
     '{{no-results-title}}': 'No Results Found',
     '{{of}}': 'Of',
+    '{{page}}': 'Page',
+    '{{prev}}': 'Prev',
     '{{previous-month}}': 'Previous month',
+    '{{previous-page}}': 'Previous Page',
     '{{results}}': 'Results',
     '{{search}}': 'Search',
     '{{show-all}}': 'Show all',
@@ -73,6 +78,7 @@ export default async function init(el) {
     cardsPerPage: 12,
     ietf: config.locale.ietf,
     collectionTags: '"caas:adobe-partners/collections/announcements"',
+    pagination: 'load-more',
     isArchive,
   };
